Added testData file.
Joseph, please have a look at this and add more to the file.
There are many edge cases that you think of and that you try with on the fly, but the more we can get them written down, the easier it will be in the future.
If you follow the format, I can easely convert them to real test harness files.

For example we tried with colon in a middle of a number, what should it be if the colon is at the start?
What if the colon is swapped for a full stop?

NVDA unicode braille addon is your friend.