en-ueb-g2.ctb after changes
This adds the work I did for the en-us-g2.ctb for after.  The ueb rules are the same for af so I just moved it over.  It correctly translates all the after words but a couple that have the en contraction because we have not fixed that contration yet..  This also gets us down to only 3100 bactrnaslation erros down from 31,000 a couple days ago.  We still have a lot of work to do on low word group signs but I will add a couple a day.