en-ueb-g2 and en-ueb-chardef fixed contractions
This commit fixes the following problems found when comparing the duxbury UEB translations to ours.  These were the only forward translation problems found:

abeyance - leaving out dots 46 before the e in the ance contraction.
ability --BB spells out ity. It should be contracted as dots 56y.
abound - BB is missing dots 46 before the d in the ound contraction.
abrasion - BB is missing dots 46 before the n in the sion contraction
abundance - BB is missing dots 46 before the e in the ance contraction
fatherless - BB is missing dots 46 before the s in the less contraction

After this commit there is one back translation issue with abound but that should be fixed in another commit.