Add test_exhaustive.py
which gives you a nice and easy way to test a lot of different
combinations of tables rules, and makes for readable tests that can be
used as documentation too.

Written in Python because I wanted to have a prototype quickly. Could
possibly be integrated in the YAML framework later.

I've written it mainly with testing and documenting the new expected
behavior of capitalization/emphasis opcodes in mind, but can imagine it
to be useful for many other things.

The test approach is based on table "templates" and table
"patterns". See the new_caps_behavior test for an example of how this
works.

The new_caps_behavior test is just a start and should be extended.