Fix back translation attribute tests for corrections
Back translation attribute tests for corrections are handled as though
they were for passes other than for the first character

You'll see two tests in that code - one for the first character, and one for
the rest. The first character passes in the variable m (0 for corrections, 1
for non-corrections) and the other passes in a hard-coded 1 (non-correctionsw).
This commit makes both tests be the same so that all characters are treated the
same. The difference is whether or not the characters are matched as characters
or as dots.