Back translation using pass rules that match characters wasn't working
The code used to use a static variable which was correct when the
length was 0 but which wasn't correct when non-zero lengths were being
passed to it. This change uses the actual length from the actual rule
rather than a hopefully correct length in a global variable. I could
have fixed the global variable, but why when it was easy to just make
the code accurate by using the real value.

Without this fix, the array of test instructions was starting at the
beginning of the characters to be matched (0) rather than where they
actually start (charslen).