Always update posMapping in for_updatePositions
Previously it was only being updated when inputPositions and
outputPositions are not both NULL, while in other functions it was
updated unconditionally.

Note that posMapping is not only used for computing inputPositions and
outputPositions, but also for updating typebuf after pass 0. However
for_updatePositions is not used in pass 0 which is why emphasis was
working fine.