Add proper support for cursor position testing in checkyaml
The cursor position can either be specified as a single number. In
that case you simply want to provide the initial cursor position for
the translation. But you if you are interested in the returned cursor
position after the translation you can provide a tuple where the first
number is the position before and the second number is the position
after the translation.

The YAML files are fixed and should now contain the original cursor
position information as in the json tests.