Disable case sensitivity checks when capitalization is not enabled
When a table does not define any capitalization opcodes we would
expect liblouis to allow contractions regardless of any
capitalization, be it at the beginning or in the middle of a word.

This patch introduces a new attribute of the table struct
`usesCapitalization` that is set to 1 if the table contains any
capitalization opcodes. The translation then only does capitalization
emphasis handling if this attribute is set to 1. Also checks whether a
rule can apply across capitalization changes are also only done for
tables that have capitalization opcodes.

Fixes #507