Remove the code that implements the "modern" pass compiler
This code was to replace the "assembler" like pass2 opcode language.
Unfortunately it was never documented and hence never used in the
wild. So it just sat around for years waiting for some malicious
tables to come along and make it over-flow some buffers.

So while this might technicaly be a more thought out language to
express your pass2 rules it never caught on and probably never will.
It is just more code to maintain that has never been battle tested.

If we'd decide to use this code then there is not only the problem of
making it work (it has never been properly tested) but also to migrate
all the existing tables.

So IMHO the best code of action is to just remove this code. Less code
means a smaller attack surface for malicious tables.