adding a few remaining exceptions to the table, correcting some punctuation
The table was not prefixing dot 5 to the question mark symbol, instead it was prefixing it to the : symbol. Yaml test have also been added to rectify this, which has been fixed in the table (last two tests). 
the remaining xfails are because:
- dot 5 should be added before any word which contains x, w or k in it and it shall not be contracted at all.
I would believe that is the only multipass scenario for forward translation.
For some reason, checkyaml throws up an error saying that in line 1, character 1 there is no yaml event. Not sure what caused this though. But the rest is functional and working.