en-ueb-g2.ctb: Words which do not use THE
Words containing the combinations THEAD, THEART, -theim etc. should not use THE because the H is aspirated as part of a compound word (RUEB 10.11.2)
Delete some specific rules and add general rules.
Generalised combination ESTHE[ST][EI] to use THE in preference to ST.
Note: May need to revisit if in combination STHE, THE should be the default.
298ff: Added general rules, deleted specific rules, added exceptions.