Separate the Dockerfile to build mingw binaries into two files
For a while I thought that running wine under win32 required a virtual
X server[1], so I wanted to get that whole mess into a separate file.
For that reason there are two Dockerfiles now.

However I now do not think a X server is needed anymore, but it still
seems cleaner to keep the two files apart. The win32 which is more
involved builds on top of the win32 image.

The win64 docker build works flawlessly, runs the test suite and all.

The win32 docker build is not quite so slick. For a while wine would
segfault, but now the build doesn't even find libyaml.

Footnotes:
[1]  https://medium.com/faun/running-windows-app-headless-in-docker-15ff008f2f16