Make the 32bit image self-contained
It doesn't really make that much sense to share the same base image as
the 64bit version. Yes there are a few things in common. But it would
mean that we'd have to uninstall a few packages (such as wine64). Also
the 32bit image needs to add the 32-bit architecture and add packages
that way.

Let's just say the two images differ in enough ways to keep them
separate.