Increase LETSIGNSIZE to 256
and at the same time decrease LETSIGNSIZE for noletsignbefore and
noletsignafter (by introducing two more constants LETSIGNBEFORESIZE
and LETSIGNAFTERSIZE.

This will allow for 256 letsign characters while keeping the size of
the table the same because we decrease the size of the noLetsignBefore
and noLetsignAfter array.

This fixes #529 in a simplistic but good enough way and helps with #921.