Merge the attribute and class opcodes
See issue https://github.com/liblouis/liblouis/issues/472

Also check that an attribute name contains only letters and digits,
like it is stated in the documentation.

Also make a number of attribute names reserved so that we have the
option to give these attributes special behavior in the
future (numericnocontchars, numericmodechars, capsmodechars, etc).