New "noempclass" opcode that is a generalization of "space" for emphasis handling
This changes the behavior a very little bit for tables that do not
declare capsmodechars or emphmodechars, and a bit more substantially
for tables that do. Tables and tests have been updated accordingly.

The code has also been restructured in a way that makes it easier to
see how "caps" differs from "emph" and how the various character
attributes and the capsmodechars, emphmodechars and noemphchars rules
affect the behavior. Extracting out this logic into a few simple
functions made both that logic and the rest of the code easier to
follow.

See https://github.com/liblouis/liblouis/issues/905