Simplify the texinfo code by adding two more macros
There is a macro to generate the text "foo opcode" with a reference.
Quite often though you want to have a reference to an opcode without
the preceding "foo opcode" text. For that there are two new macros
`opref` and `pxopref`. They are just like `ref` and `pxref` but they
make sure the reference is correct (by making it "foo opcode") and
they render the opcode inside a code