Do not upload js artifacts to release page
PRs do not have enough credentials to upload to the (snapshot) release
page. But we'd still like to run the emcripten build on PRs. So for
now just upload the artifacts

Do the upload to the snapshot release in a separate job and workflow
that is only triggered on the master branch.