Updated the Danish tables to the new 2022 Danish braille standard
The Danish braille code has been changed significantly, and is now
called "Den Danske punktskrift 2022" or DDP2022. The standard was
originally released in 2021, but was then changed again and is now
officially 2022.

Changes include:

- Limited contraction level (grade 2l or 1.5) is no longer part of the
  Danish braille standard 2022.
- There is no longer any need for "literary" tables. In Danish
  braille, capital letters are now always indicated, and steps have been
  taken in the standard to ensure better back-translation.

The Danish Braille committee recommends implementing the new standard
immediately as the default Danish Braille, but keeping the option to
use the old 1993 standard for at least a year. So, the old tables have
been renamed to include _1993 in their filenames.

See https://github.com/liblouis/liblouis/pull/1247.