Fix the bidi tests for mixed number and words
When a word contains numbers the word following the number should be
contracted. So the 'ch' in '8fach' and the 'st' in '32stel' should be
contracted.

For some reason however the backtranstion of 'st' fails. This bug is
now exposed as the forward translation correctly does contraction.
That is why I had to set the test with '32stel' to xfail.