Set doingMiltind to the actual number of rules in the multind
You can define any number of opcodes in the multind opcode. When
compiling these are stored in the multindrule->charsdots.

Now when backtranslating we presumably should go though all these
rules before moving on to the next dot pattern.

That's why we set doingMultind to the number of opcodes in the multind
opcode.