Set the token permissions to contents: read
This fixes a number of code scanning alerts such as
https://github.com/liblouis/liblouis/security/code-scanning/46.

The fix was provided by https://app.stepsecurity.io/

Two alerts were not fixed as the stepsecurity tool did not provide the
usual fix by setting the token permission. Instead it suggested to use
a "Harden Runner". ATM I do not understand the implications of that so
I postponed this.

So for mingw.yml and for emscripten.yml the alerts remain.