Add tests from MathCAT that have two files per test: 1. The MathML input 2. The expected output (as Unicode braille chars)
Note: UEB does not have a unique answer, although there is guidance from ICEB and BANA about perferred output (although they are in conflict in some cases).

As noted in the README, there are a few instances where the expected output differs from the source due to conflicting guidance. These are explained in the README.