Updates to the Norwegian braille standard
The Norwegian braille authority has made some changes to the braille
standard used in Norway. The most notable is that we no longer want to
have braille patterns to surround URLs and email addresses, if not
present in the original document.

Additionally I have added some more tests, where some of them fails. I
want to look further into that after the upcomming release.

The attached file no-no-6dot-comp6.uti is a new file which goes in the
tables directory.