Update UEB file to clarify that the UEB tests are not limited to those examples in the RUEB rulebook
As suggested in the discussion at PR #1491, to avoid creating many different test files in response
to individual issues, UEB table-related tests will be stored in a single YAML file. Historically this
file contained only examples from the UEB rulebook, but this clarifies that the examples are not
limited to that source.