Fix kannada.cti issue #1510
see https://github.com/liblouis/liblouis/pull/1510:

1 issue reported here: https://github.com/liblouis/liblouis/pull/742

The same issue plus others reported on mailing list in thread starting here:
https://www.freelists.org/post/liblouis-liblouisxml/requesting-you-to-update-lib-louis-Kannada-language-Braille-Table

The person reporting, Srini, was unable to get in touch with the
maintainers so I put him in touch with one of them, Dipendra Manocha
of the Braille Council of India (BCI).

I offered to create a PR with the changes and tests needed if no one
else could be found to do this. Dipendra welcomed this offer.

Srini provided the mailing list with the up-to-date documentation of
the correct Kannada Braille code. Also a summary of the changes needed
to the existing table. As follows:

| Unicode: | kannada Letter: | wrong braille code: | Right Braille code: |
| -------- | --------------- | ------------------- | ------------------- |
| 0C8A     | ಊ               | 14                  | 1256                |
| 0C92     | ಒ               | 346                 | 1346                |
| 0CAE     | ಮ               | 135                 | 134                 |
| 0CE0     | ೠ               | not set             | 6,1235              |

I made the changes and created tests with test words provided by
Srini. Srini tested the table and confirmed the table was worked
correctly.

From the documentation I noted 3 other characters missing from
existing table :

| Unicode:  | kannada Letter:  | Right Braille code: |
| --------- | ---------------- | ------------------- |
| 0CBD      | ಽ                | 2                   |
| 0CB1      | ಱ                | 12456               |
| 0CDE      | ೞ                | 5-456               |

Srini said these characters are not in use.