Fix various Serbian backtranslation issues (#1523)
Added a new Serbian table for Cyrillic script. This table works the
same way as the main Serbian table when forward translating, but when
back translating it prefers using the Cyrillic script. The default
table prefers the Latin script for compatibility.

Duplication was avoided as much as possible by putting common
definitions (Latin letters, Cyrillic letters, and other common
characters) in shared table file.

Both tables are tested in a single YAML file.

There are still back-translation issues left. Some issues are
difficult to resolve due to ambiguities, e.g. the left and right
parenthesis are the same pattern.

The "grade" metadata field was removed as there are no grade
distinctions in Serbian braille.

Signed-off-by: Sekulum Forka <dictatorofturkmenistan@protonmail.com>