Updates description with more information.
Updates proposed braille for pasekh-tsvey-yudn to 356-356, which looks like tsvey yudn 245-245 but one row lower, to give the two ligatures conceptual similarity. This one was Matthew Shifrin's idea, and I think it's a great one.

Adds rafe (the line that appears above veys and fey). Thanks to Mark David and Harry Bochner for the help!

Overrides Hebrew table for the unmarked beys letter to render as 12 like the beys with dagesh, rather than rendering it as 1236 like the veys with rafe.