Allow base rule in combination with character definition rule if the dots are compatible
meaning if the real parts (dots 1-8) of the dot patterns match.

For instance, with the following rules:

   lowercase a 1
   base uppercase A a
   letter A 19

you would specify that "A" is the uppercase of "a", that both should
be represented with dot 1, and that the uppercase variant should get
an additional virtual dot 9.

The dot 9 will be added regardless of whether other capital indicators
are defined. E.g. the following rule:

   capsletter 6

would cause dot 6 to be inserted before uppercase letters, but dot 9
will be added as well.

While the additional indication of uppercase is actually redundant, it
could still be useful, e.g. to produce ASCII braille for non-braille
readers (such as parents, peers or teachers), in which case the
redundant info is helpful.