Two new tables for ancient language studies
These tables include several ancient language tables in order to
obtain braille for multiple ancient languages on a single page in
primary and secondary research materials. This includes dictionaries,
commentaries, and academic articles.

There is one variant for each of the variants of Akkadian braille (the
U.S. system and the Borger system).