Fix the metadata
The values of a metadata field (that is to be used for table
discovery) have very struct syntactic requirements. No spaces are
allowed.

So I changed the authority from '#+authority:' to '#-authority:' (note
the minus sign). Then it is no longer a metadata field which is used
in the table discovery and can contain spaces.

Correct copyright mark up and angle brackets for address