Add table with common Hebrew characters
With the helpful reviewing of Leonard and Eric.

See https://github.com/liblouis/liblouis/pull/1658.

These are just the consonants. Vowels were left out. A separate table
will be provided for the vowels in the IHBC standard. Not all tables
that include the common consonants may want to use this vowel system.