Fix metadata of Russian 6-dot computer braille table
I think we may have incorrectly concluded that this is not a computer
braille table, and tagged it "literary", based on this comment:

> This table is not a computer braille in usual sense.
> It is intended for producing printable braille representation
> of a program source text, especially when it contains
> Russian comments or literals. Use it as an alternative
> literary braille table.

However, since it is intended for program source text, calling it a
computer braille table is still more accurate than calling it a
literary table IMO, even if the text can include Russian comments.

The fact that it is a 6-dot table and has no one-to-one mappings, does
not mean it can not be a computer braille table. We also have 6-dot
computer braille tables for U.S. English, German and Portuguese.

See commit 50ad4db1 and https://github.com/liblouis/liblouis/pull/903.