Drop the API to get the loglevel and set via the environment
This provides a very simple API: You can just set the log level.

Setting the log level via the API _always_ takes precedence. The
environment is only checked once and only considered if the log level
has not been set explicitly via the API.

Presumably if you set the log level as an application you do not need
a way to query it (since you set it and it always takes precedence).

It is the responsibility of the application using the liblouis API to
provide some user configuration via the environment. For example if
NVDA wants to provide the user with the ability to set the log level
they could have an environment variable NVDA_LOGLEVEL and could pass
that to liblouis via the API call.